refactor: 모듈식 폴더 구조로 재구성
