+++ ai/app/api/v1/__init__.py
@@ -1,0 +1,4 @@
+"""API v1 Endpoints"""
+from . import personality, diary, chat
+
+__all__ = ["personality", "diary", "chat"]
