+++ ai/app/core/__init__.py
@@ -1,0 +1,4 @@
+"""Core Configuration"""
+from .config import logger, CORS_CONFIG, APP_CONFIG
+
+__all__ = ["logger", "CORS_CONFIG", "APP_CONFIG"]
