+++ ai/app/services/__init__.py
@@ -1,0 +1,4 @@
+"""Business Logic Services"""
+from .diary_service import extract_recommendations
+
+__all__ = ["extract_recommendations"]
