+++ ai/demos/__init__.py
@@ -1,0 +1,1 @@
+"""Demos and Examples"""
