--- ai/environment.yml
@@ -1,23 +1,0 @@
-name: bert-emotion
-channels:
-  - pytorch
-  - conda-forge
-  - defaults
-dependencies:
-  - python=3.9
-  - pip
-  - pytorch
-  - torchvision
-  - torchaudio
-  - pip:
-    - transformers==4.36.0
-    - torch
-    - streamlit==1.28.0
-    - pandas
-    - numpy
-    - scipy
-    - scikit-learn
-    - matplotlib
-    - seaborn
-    - plotly
-    - requests
