+++ ai/requirements.txt
@@ -1,16 +1,30 @@
-torch>=1.9.0
+# FastAPI 의존성
+fastapi==0.104.1
+uvicorn[standard]==0.24.0
+pydantic==2.5.0
+
+# AI/ML 의존성
+torch>=2.0.0
 transformers==4.36.0
-streamlit==1.28.0
+langgraph>=0.2.0
+langchain>=0.3.0
+langchain-core>=0.3.0
+
+# 데이터 처리
 pandas>=1.3.0
 numpy>=1.21.0
 scipy>=1.7.0
 scikit-learn>=1.0.0
+
+# 시각화 (Streamlit 데모용)
 matplotlib>=3.4.0
 seaborn>=0.11.0
 plotly>=5.0.0
-requests>=2.25.0
-aiohttp>=3.8.0
-python-dotenv>=0.19.0
-langgraph>=0.2.0
-langchain>=0.3.0
-langchain-core>=0.3.0
+streamlit==1.28.0
+
+# HTTP 클라이언트
+requests>=2.31.0
+aiohttp>=3.9.0
+
+# 환경 설정
+python-dotenv>=1.0.0
