--- ai/requirements_fastapi.txt
@@ -1,13 +1,0 @@
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-pydantic==2.5.0
-requests==2.31.0
-aiohttp==3.9.1
-python-dotenv==1.0.0
-
-# 기존 AI 의존성
-streamlit==1.28.1
-torch==2.1.1
-transformers==4.35.2
-langgraph==0.0.20
-langchain==0.0.335
